+++ .gitignore
@@ -1,2 +1,3 @@
-src/vendor/*
+bin/
+vendor/*
 !vendor.json
